pyproject.toml
[project]
name = "flight-price-tracker"
version = "0.2.0"
description = "Async flight price tracker with email alert"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "aiomysql",
    "redis",
    "httpx",
    "python-dotenv",
    "amadeus",
    "cryptography"
]